VulkanWrapper: fix loading driver when lib path can be resolved by dlopen
On Linux-based machines, a shared library's path may be resolved in
multiple ways, such as from LD_LIBRARY_PATH. For these OSes, use dlopen
to see if it exists. Note that trying to create a vk::DynamicLoader with
an invalid path throws an exception, and enabling exceptions is not
always possible, so we want to detect the error case beforehand.

Bug: b/177624844
Change-Id: I5f27cfe485afbac0e4f6bda14eed2809669c2510