Don't initialize memory allocations in MSan builds
This allows detecting more potential uses of uninitialized memory in
MemorySanitizer (MSan) builds.

It is the application's responsibility to clear memory allocated by
vkAllocateMemory(), or guarantee that it will be written to before being
read. SwiftShader's internal uses of memory allocated by sw::allocate()
should be MSan error free at this point (explicitly initialized where
needed), and this change helps catch regressions.

Non-MSan builds are unaffected by this change, so that any errors that
are found in the near term don't pose a security risk. A subsequent
change will remove initialization for non-MSan builds as well.

Note that allocations made by the Reactor JIT-compiler will always
remain zero-initialized, since, contrary to the usage of sw::allocate(),
they are unrelated to Vulkan API allocations or other parts of the
graphics/compute pipeline which don't guarantee initialization, and are
not under the application's control.

Bug: b/140991626
Change-Id: I2c2a6c539ff48a5e7338d6c69e3e0c20a6ff7642