Fix Win32SurfaceKHR::getSurfaceCapabilities asserting when hwnd is no longer valid
When vkGetPhysicalDeviceSurfaceCapabilitiesKHR is called, we now return
VK_ERROR_SURFACE_LOST_KHR on Windows if the window handle (hwnd) is no
longer valid.

The assert was being tripped by Chromium's compositor_unittests for test
LayerWithRealCompositorTest.BackgroundBlur. With this fix, instead of an
assert, the test now fails because eglQuerySurface fails with
EGL_BAD_SURFACE.

Bug: chromium:1174372
Change-Id: I71164c30bddaa41753472389e996cebbff7fbf77