Deduce all sampler parameters from SamplingRoutineCache::Key
This cl changes createSamplingRoutine to capture only the "device"
variable, instead of capturing all function variables. The "device"
variable is required to find the sampler state from the sampler id.

This cl also changes the parameters of the getImageSampler function
so that it uses only the minimal parameters required to produce the
sampling routine.

Bug: b/152227757
Change-Id: Id7f5e18e09f078589a1a1edc12622ed40126cd32