Obtain all sampler parameters through SamplingRoutineCache::Key
This change ensures that the descriptor state identifiers used to
perform lookups in the sampling routine cache are all that is used to
obtain the state itself which is used for specializing sampling
routine generation.

The createSamplingRoutine lambda function is made to capture only the
'device' variable, instead of allowing access to all local variables.
The device is required to obtain the sampler state from the sampler
identifier.

Bug: b/152227757
Change-Id: Id7f5e18e09f078589a1a1edc12622ed40126cd32