WIP - Add safety guard for X11
This cl adds a mechanism to that extension strings can easily be
marked as supported or unsupported base on a boolean value determined
at runtime.

It then uses this new mechanism to make VK_KHR_xlib_surface
unsupported if we haven't successfully loaded X11.

Finally, the XlibSurfaceKHR methods are now themselves guarded so
that they don't crash on a null X11 or a null display.

Bug: N/A
Change-Id: I6da3755a15e09320e5715b2036043acbb31c0c35