Remove vkGetInstanceProcAddr from exported symbols on Fuchsia
Fuchsia only allows access to the ICD's functions, through the Vulkan
Loader interface functions.

Exporting this function was copied from the android_vk_swiftshader.lds
file, but it has been removed from the latter due to causing direct
access to methods that should have an overridden implementation. Similar
issues are likely to occur on Fuchsia if we keep this function exported.

Bug: b/148822450
Change-Id: I12f2dad933abcd9dbf89518ac5714b9f936419e4