Implement VK_EXT_depth_clip_enable
Normally the vertex processor will clip geometry that's outside of the
near and far planes. However, when depthClamp is enabled, depth clipping
must be disabled. So this change also enables the extension that allows
users to specify if depth clipping is enabled or disabled.

Bug: b/185814882
Tests: dEQP-VK.clipping.clip_volume.depth_clip.*
Change-Id: Iaab31c17cac382cf55a8c50d8e7e7d4c87b55272