Add PRESUBMIT.py to SwiftShader
PRESUBMIT.py is run by git cl upload to do some rudimentary checks on a
change. Currently, this only makes sure that the files in a CL are
properly formatted, but the script is easily extendable to cover
multiple facets of a CL like running unittests, ensuring a commit
message references an issue, etc.

Bug: b/187094215
Change-Id: Icedbdcfe2fcdcc2b318738da6aaab19c2e0e69fa