Make clang-format-all.sh git aware
clang-format-all.sh will modify every file in the repository that
isn't properly formatted. However, if you only care about the
formatting of the files touched in a CL, this is over-zealous. So
the default behavior is to only modify a file list derived from git's
current set of files.

You can still use the script to format the entire repo with -a or --all

Bug: b/187094215
Change-Id: I2753c83e00d36817eddb0d8e641100240a2e8e23