Work around MemorySanitizer scalar SSE false positives
Scalar SSE instructions only use the lowest scalar of an SSE vector
register, but MemorySanitizer doesn't recognize some of them so it will
check the entire 128-bit operand for uninitialized bits.

This change makes sure the other elements of the vector gets zero-
initialized instead of leaving it undefined. This affects Round, Trunc,
Frac, Ceil, Floor, Sqrt, and RcpSqrt scalar operations.

Note that this workaround results in MemorySanitizer marking the entire
output vector to have a well-defined value, which isn't actually the
case in non-MSan builds. Fortunately, a scalar can't be cast into a
vector (unlike our 'emulated' small vectors), so we just have to make
sure to immediately extract the scalar from the intrinsic's result.

Bug: b/172238865
Change-Id: If68388e476ac9e27e2de33ddf2efab4124540c7a