Implement VK_EXT_custom_border_color
This implementation uses the custom border color as "baked" state,
instead of data stored in the descriptor. This is justified by the
assumption that only few custom border colors will be used in practice.
maxSamplerAllocationCount and maxCustomBorderColorSamplers are both
set to 4000, like many other implementations currently:
https://vulkan.gpuinfo.org/displayextensionproperty.php?name=maxCustomBorderColorSamplers

Note that infinity is deemed a valid border component value, and Reactor
intentionally asserts when trying to initialize a Float constant with
infinity, so we bit-cast to int in C++.

Bug: b/151215666
Bug: angleproject:6200
Tests: dEQP-VK.*clamp_to_border_custom*
Change-Id: I22608a80f50194bbac66cf515b0a9249b67b5f00