Marl: Suppress MemorySanitizer false-positive
Worker::current is a static thread_local which gets intialized to null,
which gets checked for in Scheduler::Fiber::current(). Unfortunately
MemorySanitizer does not observe the initialization when Marl is used
within a shared library, when the Linux dynamic loader is not
instrumented and its TLS initialization is not otherwise intercepted and
handled correctly.

This change decorates Scheduler::Fiber::current() with
__attribute__((no_sanitize_memory)) to suppress instrumenting it with
use-of-uninitialized-data checks.

Bug: chromium:1211047