Disable -warn-stack-size for LLVM older than version 12
This command line option is getting replaced with a module attribute in
https://reviews.llvm.org/D103928. Temporarily disable it for LLVM
versions newer than 11. Once LLVM 12 is released we should be able to
use the attribute instead to retain this functionality.

Note that upstream SwiftShader still uses LLVM 10, and the primary usage
of this functionality is to prevent stack overflow in Chrome's fuzzers,
which is not affected by this temporary change.

Bug: b/191193823
Change-Id: I2edb0dc4df5d434540741bcb3007539311eb958a