Enable the -Wdeprecated-copy compiler warning
Previous occurences of this warning have been fixed. Note the warning is
also added by -Wextra so we previously explicitly disabled it with
-Wno-deprecated-copy. While removing the latter should suffice to re-
enable it, it's useful to make it explicit since support for implicit
copy constructors when a user-defined assignment operator has been
defined may be removed in the near future.

Bug: b/191417833
Change-Id: If6721ae900afd530750a7d05ccc40365924d4c25