Make UpdateAfterBind descriptor limits equal to base limits
The Vulkan 1.2.182 spec mandates that "The UpdateAfterBind descriptor
limits must each be greater than or equal to the corresponding
non-UpdateAfterBind limit."

Note that at the moment we do not advertise any of the update after bind
features (e.g. descriptorBindingSampledImageUpdateAfterBind), but the
Vulkan Validation Layers still expect these limits to be non-0.
A descriptor pool creation flag and a descriptor binding flag which are
disallowed when the features are not enabled prevent actual update after
bind usage.

Bug: swiftshader:160
Change-Id: Icce2ba987cb67a87544a406df144dfce8026b3f6