Fix -Wundefined-var-template.
This is a cryptic error, so I'll explain.  See also
https://lists.llvm.org/pipermail/cfe-commits/Week-of-Mon-20160418/155997.html

When the compiler encounters a reference to a static member of a
templated class, it needs a corresponding definition to instantiate it.
As long as one occurs in some translation unit (TU), this is fine, but
if none occur, the resulting problem is difficult to diagnose.  So this
warning appears in every TU that does not see a definition.

In Subzero's case, the definitions of various such members occurred in
individual .cpp files, but other .cpp files referred to these members,
producing warnings.  The recommended fix -- which is applied here -- is
to move the definitions from individual .cpp files to common .h files
shared by all relevant TUs (and let the linker sort out the multiple
instances, as it does with other template instantiations).

Bug: chromium:604888
Change-Id: I63b58ecdf956ff264e6d25738684b513f05b268b