Implement Reactor basic constant folding for Int
Constant folding and propagation is achieved at the Reactor level by
storing an Immediate value in RValue<Int> and Int, and using it to avoid
the creation of backend Value objects for constants. Currently only
addition and multiplication is folded, which suffices to eliminate all
overhead from dynamic indexing using Int variables when they are
actually static (b/146166966).

Provide documentation for this design, as well for the prior
'materialization' optimization (b/129356087).

Modified the 'Sample' unit test to include a case of an RValue<Int>
being constructed from a Reference<Int>.

Bug: b/155302798
Change-Id: I993465f970a03ed333410bebe8200f14796def7a