Regres: Increase MaxProcMemory to 6 GiB
Increasing the maxComputeWorkGroupInvocations to 256 revealed that we're
consuming a lot of memory with the Subzero JIT for compute shaders that
contain control barriers, due to implementing them with the use of
fibers which each have a 1 MiB stack.

Increasing Regres's maximum memory consumption per process keeps the
tests passing while we work on a more long-term solution.

Bug: b/192475710
Change-Id: I027d97d9f28d9749628fc008fe956fd436c77562