Remove mips-specific arch flags
Most of these flags are already in the default //build setup
in build/config/compiler/BUILD.gn, config("compiler_cpu_abi").
`-mxgot` isn't -- if that flag is needed, it should probably
be in the upstream gn build files. If it's needed just for
swiftshader we can add it back with a comment that says why
the flag is swiftshader-specific.