Rename Subzero traits Operand to AsmOperand
Ice::Operand is a base class representing high-level operands.
X86Operand is a subclass of it which is an x86-specific operand, but
still part of the high-level IR.

The traits Operand has no direct relationship to that hierarchy, and
just holds low-level x86 encoding parameters related to an operand.
Since it's mostly used by the AssemblerX8632[64|32] class, it is being
renamed to AsmOperand to disambiguate it from Ice::Operand.

This is a step towards making the X86 backends traits free.

Bug: b/192890685
Change-Id: I588ca87b73c6502522bb09dd41a9ce73446ea601