Rename Subzero Address to AsmAddress
For consistency with AsmOperand, which is the base class of Address,
this change renames the latter to AsmAddress. It also makes it clearer
when we're dealing with an (x86) assembler-level parameter, vs. a high-
level IR parameter.

stackVarToAsmOperand was renamed to stackVarToAsmAddress since it
returns the address subclass. Redundant namespace and traits member
qualifiers were removed.

Bug: b/192890685
Change-Id: I65f3ea4d490f44794e8a7eeb46391cbfd40d54cf