Fix RequiredStackAlignment initialization
sizeof(Traits::WordType) evaluates to obtaining the size of the numeric
value of the IceType_i32 or IceType_i64 enum. Because enums default to
being constants with an underlying type of 'int', this becomes 4 for
both the x86-32 and x86-64 backends, while 8 was intended for the
latter.

A misaligned stack may have a significant performance impact.

Note that an even greater stack alignment might be required, as both the
Microsoft and System V x64 ABI require 16-byte alignment. This change
only addresses the obvious semantics bug.

Bug: b/193550986
Change-Id: If33c0a10d73c50bfb1bbc25f16acb97b3575e27a