Increase Subzero x86-64 stack alignment to 16
The SystemV x64 ABI requires that "The stack is 16-byte aligned just
before the call instruction is called."

The Microsoft x64 ABI states that "The stack will always be maintained
16-byte aligned, except within the prolog (for example, after the return
address is pushed), and except where indicated in Function Types for a
certain class of frame functions." The latter refers to functions which
do not call other functions but are not leaf functions (i.e. they may be
called externally, or allocate additional stack space).

This change fixes the calling convention and improves performance for
spilled 128-bit vectors.

Bug: b/193550986
Change-Id: If26600b701359da996b288a86ec92e07132913b2