Backport LLVM fix for mallinfo deprecation
This backports the following LLVM commit with a few
SwiftShader specific modifications, like testing for
__GLIBC__ because SwiftShader is used within Chrome
and built via GN bypassing cmake and not doing any
build time autodetection.

Bug: b/192875071
Test: Local builds on ChromeOS with glibc 2.32 & 2.33

Original commit message:

Fix deprecated usage of `mallinfo`

glibc deprecates `mallinfo` in the latest version of 2.33. This patch replaces the usage of `mallinfo` with the new `mallinfo2` when it's available.

Reviewed By: lattner

Differential Revision: https://reviews.llvm.org/D96359