Only process active fragment samples
This change refactors the handling of which samples we should iterate
over, by putting their indices into a container object. This enables the
use of range-based for loops to have a more elegant syntax and avoid the
confusion around sampleId -1 having special meaning and the begin and
end iteration markers still requiring checking the sample mask on each
loop.

Bug: b/194521425
Change-Id: Ib6fbbb3e89c3a5501311ebd81859608df44d1bd0