Fix computing anisotropy factor for single-level mipmap
When anisotropic filtering is enabled, we need to compute the anisotropy
factor. This computation is part of the LOD computation, so don't skip
it even when there's only a single mipmap level.

The computation of the LOD and anisotropy are a bit intertwined because
we need to adjust the LOD according to the anisotropy, but we could
still skip a few calculations in the single-level mipmap case. This
change does not perform this micro-optimization yet.

Bug: swiftshader:163
Bug: b/151263485
Change-Id: Ib250cfd9ec2d784761887c1c1f58c4127fa6d92e