Regres sandboxes dEQP [WIP]
Regres runs on corporate machines and executes unchecked code. While
regres currently only runs CLs from internal google accounts, it's in
our best interest to restrict its access to the bare-minimum.

Firejail is an ISE approved sandboxing tool for Linux. It uses a profile
to determine what resources the program it's running can use. Regres
creates a profile that only allows dEQP to read and write to its own
directory (necessary for amber) and read and write to the log file
specified.

This patch treats firejail as an optional dependency as it is not
available on Windows.

Bug: b/195441834
Change-Id: I1199000640e27f55f5aadff50fb97569db86c1af