Fix alignment mismatch between IOSurface and SwiftShader
IOSurface objects have their own alignment requirements which
may be stricter than SwiftShader's requirements. For example,
in local tests, the bytesPerRow property of IOSurface objects
had a required alignment of 64 bytes, which is larger than
SwiftShader's requirement, which is only for the width to be
even.

This was discovered while debugging the Clusterfuzz issue below,
but this doesn't fix the memory corruption, only the stride issue.

Bug: chromium:1232377
Change-Id: I31852bc8b5b902142d0875cb88e5fd5848425732