Deprecate libEGL and libGLESv2 CMake build targets
This change marks the OpenGL ES build targets for CMake as deprecated.
The legacy code is unmaintained as is being replaced by "SwANGLE", the
combination of ANGLE and SwiftShader's Vulkan implementation, which
offers conformance and performance benefits.

Android is fully switched to SwANGLE at this point, and Chromium GN
builds are also largely switched over. Remaining usage of the CMake
based builds outside of our team is lesser known, and so this change
serves as a final notification of our plans to remove the legacy code.
For the time being, the deprecated build targets can still be used, but
we urge people to use SwANGLE instead at their earliest convenience.

Building SwANGLE is documented in docs/ANGLE.md

Any issues to use SwANGLE can be reported at http://g.co/swiftshaderbugs

Bug: b/147516027
Change-Id: Ic22616af3324368346f0cfee9f3c9b77e951410e