Don't allocate an external memory object when there's no external memory
The current logic was very confusing, since the presence of
the "external" member in DeviceMemory did not mean that the
memory was external. The DeviceMemoryHostExternalBase class
represented non external memory. This CL deletes this class
and handles the potential null "external" within DeviceMemory
member functions. A null "external" member behaves the same
way as the DeviceMemoryHostExternalBase did, so there should
be no behavior change in this CL, but checking for the
"external" member in DeviceMemory will now tell us whether
or not the memory is external, so
DeviceMemory::hasExternalMemory() should now work as intended.

Bug: chromium:1232377
Change-Id: Ia739532772b024fecdff3c9a0200cf3e22b6fe12