Fix Subzero Optimizer run time variability
std::unordered_map<> can have greatly varying performance when using a
pointer as the key value type. That's because it allocates 'buckets' for
storing values corresponding with each key's hash value. When the
buckets get too full, more buckets are allocated and elements get
rehashed, which is a costly operation. When that happens, and how often,
is unpredictable when using object pointers as keys.

This was the case for Optimizer::optimizeSingleBasicBlockLoadsStores()
optimization pass for Subzero. This change replaces the use of the
pointer to InstAlloca instructions, which represent stack variables,
with the (unique) index of the alloca's destination argument (i.e. the
address of the allocated variable), which is of integer type.

This eliminates dramatic differences in run time, for identical input.
Specifically this makes it feasible again to reliably determine whether
a change is an optimization or a performance regression, as part of our
CI testing with Regres.

Fixes: b/193550986
Change-Id: Idf491082e4e2ffb83b2ff3e97fff2f190b0534ff