Remove global constructor from BC6
We previously changed BC6 to use an array instead of a std::vector. I
thought this would eliminate the use of a global constructor. However,
the objects within the array actually still invoke global constructors.

So I've returned to using an earlier plan we had of using a function static
table. The data is only constructed when a user tries to use BC6, and
it's stil only constructed once.

Bug: b/174782098
Tests: dEQP-VK.*bc6*
Change-Id: Ie987f0faa6239a5a2917713471d6de13897112b7