Minimize usage of global constructors in Reactor
A few const bools are easily replaced with constexprs. Other global
constructors are left as is for now.

llvm_shutdown_obj is necessary for cleaning up LLVM at program exit and
is necessarily used every time SwiftShader is invoked.

We could replace CPUID::SSE4_1 with a call to a singleton function.
However, I believe this increases code complexity for little benefit.
Because it's a bool, there's no deconstructor associated with it, it's
cleaned up when global memory is deallocated.

Bug: b/175073772
Change-Id: I25635fda0878e30668cbfc6982877b5d2d924789