Remove erroneous ASSERT for format lists
VK_KHR_image_format_list is supported by SwiftShader, yet an ASSERT was
left in to guard against it being used. Nothing needs to replace the
ASSERT as SwiftShader does not benefit from knowing which formats are
used, and it's the job of a validation layer to ensure that a user isn't
using a format they said they wouldn't.

Bug: b/198327618
Change-Id: I6c8133f8fe57161fb2ce4757023861b052639510