Refactor VkMemoryAllocateInfo parsing
The VkMemoryAllocateInfo structure, when used for
external memory, can contain various structures in
its pNext pointer. Previously, VkMemoryAllocateInfo
is parsed by external memory subclass. This CL
changes that to have a single function which parses
the VkMemoryAllocateInfo, with all potential users
of the VkMemoryAllocateInfo using a new structure,
DeviceMemory::ParsedAllocationInfo, which contains
the result of the parsing, which makes the code much
more readable. It also makes is easier to check if
an unexpected pNext structure is used, because the
parsing only happens in a single function, as opposed
to being spread out across multiple class functions.

Bug: b/199306480
Change-Id: I88598a3c6f1f2533f40cf1e1e82405fccf0771d2