Remove vulkan_fuchsia_extras.h from vulkan.h
This change ensures our vulkan.h copy matches that of the official
version we're using (currently 1.2.189). It avoids having to add the
include of vulkan_fuchsia_extras.h when updating it. Code that uses
definitions from this header now include it explicitly.

The contents of the vulkan_fuchsia_extras.h header should be upstreamed
to vulkan/vulkan_fuchsia.h.

Bug: b/143197472
Change-Id: Id67bfa9902dea91470e0c75623d16e86f1f5ba2a